// foxit_sdk/src/models/baseEnvelopefromTemplate.ts
export interface BaseEnvelopefromTemplate {
  /** If this value is not provided, then the document(s) folder name is kept same as the template(s) name(s). */
  folderName?: string;
  templateIds: number[];
  fields?: unknown[];
  allowAdvancedEmailValidation?: boolean;
  parties: Party[];
  folderPassword?: string;
  signInSequence?: boolean;
  inPersonEnable?: boolean;
  sendNow?: boolean;
  signSuccessUrlAllParties?: boolean;
  createEmbeddedSendingSession?: boolean;
  fixRecipientParties?: boolean;
  fixDocuments?: boolean;
  sendSuccessUrl?: string;
  sendErrorUrl?: string;
  hideSignerSelectOption?: boolean;
  hideSignerActions?: boolean;
  hideSenderName?: boolean;
  hideSendButton?: boolean;
  hideFolderName?: boolean;
  hideDocumentsName?: boolean;
  hideAddMeButton?: boolean;
  hideAddNewButton?: boolean;
  hideAddGroupButton?: boolean;
  createEmbeddedSigningSession?: boolean;
  createEmbeddedSigningSessionForAllParties?: boolean;
  embeddedSignersEmailIds?: string[];
  signSuccessUrl?: string;
  signDeclineUrl?: string;
  signLaterUrl?: string;
  signErrorUrl?: string;
  hideNextRequiredFieldBtn?: boolean;
  themeColor?: string;
  hideDeclineToSign?: boolean;
  hideMoreAction?: boolean;
  hideAddPartiesOption?: boolean;
  sessionExpire?: boolean;
  expiry?: number;
  senderEmail?: string;
  allowSendNowAndEmbeddedSigningSession?: boolean;
}